viết hoa các components, xóa đi các component dư thừa
