+++ src/components/Commons/SectionBreadcrumb.js
@@ -6,7 +6,6 @@ function SectionBreadcrumb() {
                 <div className="row">
                     <div className="col-lg-12 text-center">
                         <div className="breadcrumb__text">
-                            {/*<h2>Vegetable’s Package</h2>*/}
                             <div className="breadcrumb__option">
                                 <a href="">Trang chủ</a>
                                 <a href="">Chi tiết sản phẩm</a>
