+++ src/components/SectionHero.js
@@ -1,20 +1,20 @@
 
-import Categories from "./categories"
-import SearchBar from "./search_bar"
-import Phone from "./Commons/phone";
-import Banner from "./banner"
+import AllCategories from "./AllCategories"
+import SearchBar from "./SearchBar"
+import Phone from "./Commons/Phone";
+import Banner from "./Banner"
+
 function SectionHero() {
     return (
         <section className="hero">
             <div className="container">
                 <div className="row">
                     <div className="col-lg-3">
-                        <Categories/>
+                        <AllCategories/>
                     </div>
                     <div className="col-lg-9">
                         <div className="hero__search">
                             <SearchBar/>
-                            {/*<Phone/>*/}
                         </div>
                         <Banner/>
                     </div>
