--- src/components/category_item.js
@@ -1,13 +1,0 @@
-
-
-function CategoryItem() {
-    return (
-        <div className="col-lg-3">
-            <div className="categories__item set-bg" data-setbg="">
-                <h5><a href="#">Nón Lá</a></h5>
-            </div>
-        </div>
-    )
-}
-
-export default CategoryItem;
