+++ src/router/router.js
@@ -1,6 +1,6 @@
 import {createBrowserRouter} from "react-router-dom";
 import App from "../components/App";
-import {ListProducts} from "../components/list_products";
+import {ListProducts} from "../components/List_products";
 
 export const router = createBrowserRouter([
     {
